+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/Validator.kt
@@ -1,38 +1,47 @@
 package at.asitplus.wallet.lib.agent
 
-import at.asitplus.KmmResult
-import at.asitplus.wallet.lib.iso.DeviceResponse
-import at.asitplus.wallet.lib.iso.Document
-import at.asitplus.wallet.lib.iso.IssuerSigned
-import at.asitplus.wallet.lib.iso.MobileSecurityObject
 import at.asitplus.openid.TransactionDataBase64Url
 import at.asitplus.signum.indispensable.CryptoPublicKey
 import at.asitplus.signum.indispensable.cosef.CoseKey
-import at.asitplus.signum.indispensable.cosef.io.Base16Strict
-import at.asitplus.signum.indispensable.cosef.io.coseCompliantSerializer
 import at.asitplus.signum.indispensable.josef.JwsSigned
 import at.asitplus.wallet.lib.DefaultZlibService
 import at.asitplus.wallet.lib.ZlibService
 import at.asitplus.wallet.lib.agent.Verifier.VerifyCredentialResult
-import at.asitplus.wallet.lib.agent.Verifier.VerifyCredentialResult.*
 import at.asitplus.wallet.lib.agent.Verifier.VerifyPresentationResult
-import at.asitplus.wallet.lib.agent.validation.*
+import at.asitplus.wallet.lib.agent.validation.CredentialFreshnessSummary
+import at.asitplus.wallet.lib.agent.validation.CredentialTimelinessValidator
+import at.asitplus.wallet.lib.agent.validation.StatusListTokenResolver
+import at.asitplus.wallet.lib.agent.validation.TokenStatusResolver
+import at.asitplus.wallet.lib.agent.validation.TokenStatusResolverImpl
+import at.asitplus.wallet.lib.agent.validation.TokenStatusResolverNoop
+import at.asitplus.wallet.lib.agent.validation.TokenStatusValidator
+import at.asitplus.wallet.lib.agent.validation.invoke
 import at.asitplus.wallet.lib.agent.validation.mdoc.MdocInputValidator
 import at.asitplus.wallet.lib.agent.validation.sdJwt.SdJwtInputValidator
-import at.asitplus.wallet.lib.agent.validation.vcJws.VcJwsInputValidationResult
+import at.asitplus.wallet.lib.agent.validation.toTokenStatusValidator
 import at.asitplus.wallet.lib.agent.validation.vcJws.VcJwsInputValidator
 import at.asitplus.wallet.lib.cbor.VerifyCoseSignature
 import at.asitplus.wallet.lib.cbor.VerifyCoseSignatureFun
 import at.asitplus.wallet.lib.cbor.VerifyCoseSignatureWithKey
 import at.asitplus.wallet.lib.cbor.VerifyCoseSignatureWithKeyFun
-import at.asitplus.wallet.lib.data.*
+import at.asitplus.wallet.lib.data.IsoDocumentParsed
+import at.asitplus.wallet.lib.data.VerifiableCredentialJws
+import at.asitplus.wallet.lib.data.VerifiableCredentialSdJwt
+import at.asitplus.wallet.lib.data.VerifiablePresentationJws
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.StatusListTokenPayload
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus
-import at.asitplus.wallet.lib.jws.*
-import io.github.aakira.napier.Napier
-import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
+import at.asitplus.wallet.lib.iso.DeviceResponse
+import at.asitplus.wallet.lib.iso.Document
+import at.asitplus.wallet.lib.iso.IssuerSigned
+import at.asitplus.wallet.lib.iso.MobileSecurityObject
+import at.asitplus.wallet.lib.jws.SdJwtSigned
+import at.asitplus.wallet.lib.jws.VerifyJwsObject
+import at.asitplus.wallet.lib.jws.VerifyJwsObjectFun
+import at.asitplus.wallet.lib.jws.VerifyJwsSignature
+import at.asitplus.wallet.lib.jws.VerifyJwsSignatureFun
+import at.asitplus.wallet.lib.jws.VerifyJwsSignatureWithCnf
+import at.asitplus.wallet.lib.jws.VerifyJwsSignatureWithCnfFun
 import kotlinx.datetime.Clock
-import kotlinx.serialization.encodeToByteArray
 import kotlin.coroutines.cancellation.CancellationException
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.seconds
@@ -43,41 +52,40 @@ import kotlin.time.Duration.Companion.seconds
  * Does verify the revocation status of the data (when a status information is encoded in the credential).
  */
 class Validator(
+    @Deprecated("Has been moved to [ValidatorSdJwt], [ValidatorVcJws], [ValidatorMdoc]")
     private val verifySignature: VerifySignatureFun = VerifySignature(),
+    @Deprecated("Has been moved to [ValidatorSdJwt], [ValidatorVcJws]")
     private val verifyJwsSignature: VerifyJwsSignatureFun = VerifyJwsSignature(verifySignature),
+    @Deprecated("Has been moved to [ValidatorSdJwt], [ValidatorVcJws]")
     private val verifyJwsObject: VerifyJwsObjectFun = VerifyJwsObject(verifyJwsSignature),
+    @Deprecated("Has been moved to [ValidatorSdJwt]")
     private val verifyJwsSignatureWithCnf: VerifyJwsSignatureWithCnfFun = VerifyJwsSignatureWithCnf(verifyJwsSignature),
+    @Deprecated("Has been moved to [ValidatorMdoc]")
     private val verifyCoseSignature: VerifyCoseSignatureFun<StatusListTokenPayload> = VerifyCoseSignature(),
+    @Deprecated("Has been moved to [ValidatorMdoc]")
     private val verifyCoseSignatureWithKey: VerifyCoseSignatureWithKeyFun<MobileSecurityObject> =
         VerifyCoseSignatureWithKey(verifySignature),
-    /** Toggles whether transaction data should be verified if present. */
+    @Deprecated("Has been moved to [ValidatorSdJwt]")
     private val verifyTransactionData: Boolean = true,
-    /** Structure / Integrity / Semantics validator. */
+    @Deprecated("Has been moved to [ValidatorVcJws]")
     private val vcJwsInputValidator: VcJwsInputValidator =
         VcJwsInputValidator(verifyJwsObject = verifyJwsObject),
-    /** Structure / Integrity / Semantics validator. */
+    @Deprecated("Has been moved to [ValidatorSdJwt]")
     private val sdJwtInputValidator: SdJwtInputValidator =
         SdJwtInputValidator(verifyJwsObject = verifyJwsObject),
-    /** Structure / Integrity / Semantics validator. */
+    @Deprecated("Has been moved to [ValidatorMdoc]")
     private val mdocInputValidator: MdocInputValidator =
         MdocInputValidator(verifyCoseSignatureWithKey = verifyCoseSignatureWithKey),
-    @Deprecated("Use parameter in `CredentialTimelinessValidator` instead")
+    @Deprecated("Use parameter in [CredentialTimelinessValidator] instead")
     private val timeLeeway: Duration = 300.seconds,
+    @Deprecated("Use parameter in [tokenStatusResolver], [credentialTimelinessValidator] instead")
     private val clock: Clock = Clock.System,
+    @Deprecated("Use parameter in [tokenStatusResolver] instead")
     private val zlibService: ZlibService = DefaultZlibService(),
+    @Deprecated("Use [TokenStatusResolverImpl] for [tokenStatusResolver] instead")
     private val resolveStatusListToken: StatusListTokenResolver? = null,
-    /**
-     * The function [tokenStatusResolver] should check the status mechanisms in a given status claim in order to
-     * extract the token status.
-     */
-    private val tokenStatusResolver: TokenStatusResolver = resolveStatusListToken?.toTokenStatusResolver(
-        verifyJwsObjectIntegrity = verifyJwsObject,
-        zlibService = zlibService,
-        verifyCoseSignature = verifyCoseSignature,
-        clock = clock,
-    ) ?: TokenStatusResolver {
-        KmmResult.success(TokenStatus.Valid)
-    },
+    /** Clients may use [TokenStatusResolverImpl]. */
+    private val tokenStatusResolver: TokenStatusResolver = TokenStatusResolverNoop,
     private val acceptedTokenStatuses: Set<TokenStatus> = setOf(TokenStatus.Valid),
     private val tokenStatusValidator: TokenStatusValidator =
         tokenStatusResolver.toTokenStatusValidator(acceptedTokenStatuses),
@@ -120,6 +128,7 @@ class Validator(
     internal suspend fun checkRevocationStatus(sdJwt: VerifiableCredentialSdJwt) = tokenStatusValidator(sdJwt)
     internal suspend fun checkRevocationStatus(vcJws: VerifiableCredentialJws) = tokenStatusValidator(vcJws)
 
+    @Suppress("DEPRECATION")
     private val validatorVcJws = ValidatorVcJws(
         verifySignature = verifySignature,
         verifyJwsSignature = verifyJwsSignature,
@@ -148,10 +157,12 @@ class Validator(
         publicKey: CryptoPublicKey?,
     ): VerifyCredentialResult = validatorVcJws.verifyVcJws(input, publicKey)
 
+    @Suppress("DEPRECATION")
     private val validatorSdJwt = ValidatorSdJwt(
         verifySignature = verifySignature,
         verifyJwsSignature = verifyJwsSignature,
         verifyJwsObject = verifyJwsObject,
+        verifyJwsSignatureWithCnf = verifyJwsSignatureWithCnf,
         verifyTransactionData = verifyTransactionData,
         sdJwtInputValidator = sdJwtInputValidator,
         validator = this
@@ -171,6 +182,7 @@ class Validator(
         publicKey: CryptoPublicKey?,
     ): VerifyCredentialResult = validatorSdJwt.verifySdJwt(sdJwtSigned, publicKey)
 
+    @Suppress("DEPRECATION")
     private val validatorMdoc = ValidatorMdoc(
         verifySignature = verifySignature,
         verifyCoseSignatureWithKey = verifyCoseSignatureWithKey,